alterando os codigos para o tipo Long

+++ controlePedidos/src/main/java/com/br/chocolatePontoVirgula/model/dto/ItemPedidoDTO.java
@@ -11,8 +11,8 @@ public class ItemPedidoDTO {
     private int codigo;
     private int quantidade;
     private Double valorTotal;
-    private int codigoPedido;
-    private int codigoProduto;
+    private Long codigoPedido;
+    private Long codigoProduto;
 
     public ItemPedidoDTO(ItemPedido itemPedido) {
         this.codigo = itemPedido.getCodigo();
